push_key [nfc]: Move generatePushKey etc here from PushDeviceManager
