+++ lib/model/push_device.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:math';
 
 import 'package:collection/collection.dart';
 import 'package:drift/drift.dart' as drift;
@@ -81,26 +80,6 @@ class PushDeviceManager extends PerAccountStoreBase {
         }
     }
   }
-
-  /// Generate a suitable value to pass as `pushKeyId` to [registerPushDevice].
-  static int generatePushKeyId() {
-    final rand = Random.secure();
-    return rand.nextInt(1 << 32);
-  }
-
-  /// Generate a suitable value to pass as `pushKey` to [registerPushDevice].
-  static Uint8List generatePushKey() {
-    final rand = Random.secure();
-    return Uint8List.fromList([
-      pushKeyTagSecretbox,
-      ...Iterable.generate(32, (_) => rand.nextInt(1 << 8)),
-    ]);
-  }
-
-  /// The tag byte for a libsodium secretbox-based `pushKey` value.
-  ///
-  /// See API doc: https://zulip.com/api/register-push-device#parameter-push_key
-  static const pushKeyTagSecretbox = 0x31;
 
   /// Send this client's notification token to the server, now and if it changes.
   ///
