+++ lib/model/push_key.dart
@@ -1,3 +1,5 @@
+import 'dart:math';
+
 import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 
@@ -136,4 +138,24 @@ class PushKeyStore {
     await _globalPushKeys._removePushKey(pushKeyId);
     _pushKeys.remove(pushKeyId);
   }
+
+  /// Generate a suitable value to pass as `pushKeyId` to [registerPushDevice].
+  static int generatePushKeyId() {
+    final rand = Random.secure();
+    return rand.nextInt(1 << 32);
+  }
+
+  /// Generate a suitable value to pass as `pushKey` to [registerPushDevice].
+  static Uint8List generatePushKey() {
+    final rand = Random.secure();
+    return Uint8List.fromList([
+      pushKeyTagSecretbox,
+      ...Iterable.generate(32, (_) => rand.nextInt(1 << 8)),
+    ]);
+  }
+
+  /// The tag byte for a libsodium secretbox-based `pushKey` value.
+  ///
+  /// See API doc: https://zulip.com/api/register-push-device#parameter-push_key
+  static const pushKeyTagSecretbox = 0x31;
 }
